Switch to dispatcherd mgmt cmd
