+++ tools/docker-compose/supervisor.conf
@@ -4,7 +4,7 @@ minfds = 4096
 nodaemon=true
 
 [program:awx-dispatcher]
-command = awx-manage run_dispatcher
+command = awx-manage dispatcherd
 autorestart = true
 stopasgroup=true
 killasgroup=true
