+++ tools/grafana/dashboards/services_dashboard.json
@@ -102,7 +102,7 @@
             "-b",
             "provision_instance",
             "run_callback_receiver",
-            "run_dispatcher",
+            "dispatcherd",
             "run_rsyslog_configurer",
             "run_ws_heartbeat",
             "run_wsrelay",
@@ -112,7 +112,7 @@
             "-b",
             "provision_instance",
             "run_callback_receiver",
-            "run_dispatcher",
+            "dispatcherd",
             "run_rsyslog_configurer",
             "run_ws_heartbeat",
             "run_wsrelay",
